rm_pm_hal: v11.02.03

+++ pm/include/psc_soc_device.h
@@ -3,7 +3,7 @@
  *
  * Cortex-M3 (CM3) firmware for power management
  *
- * Copyright (C) 2017-2025, Texas Instruments Incorporated
+ * Copyright (C) 2017-2024, Texas Instruments Incorporated
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -50,8 +50,8 @@ struct soc_device_data {
 	u8		psc_idx;        /** PSC index */
 	u8		pd;             /** PSC powerdomain */
 #else
-	unsigned int	psc_idx : 3;
-	unsigned int	pd : 5;
+	u8		psc_idx : 3;
+	u8		pd : 5;
 #endif
 	lpsc_idx_t	mod;   /** PSC module index */
 };
